refactor(api): apply review changes
