+++ api/src/prescription/target-profile/application/target-profile-controller.js
@@ -1,6 +1,6 @@
 import { getChallengeLocale } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../domain/usecases/index.js';
-import * as frameworkWithoutSkillSerializer from '../infrastructure/serializers/jsonapi/framework-without-skills-serializer.js';
+import * as frameworkWithSkillsSerializer from '../infrastructure/serializers/jsonapi/framework-with-skills-serializer.js';
 import * as targetProfileForSpecifierSerializer from '../infrastructure/serializers/jsonapi/target-profile-for-specifier-serializer.js';
 
 const findTargetProfiles = async function (request, h, dependencies = { targetProfileForSpecifierSerializer }) {
@@ -12,12 +12,12 @@ const findTargetProfiles = async function (request, h, dependencies = { targetPr
 const findLearningContentsByOrganizationId = async function (
   request,
   _,
-  dependencies = { frameworkWithoutSkillSerializer },
+  dependencies = { frameworkWithSkillsSerializer },
 ) {
   const organizationId = request.params.organizationId;
   const locale = getChallengeLocale(request);
   const learningContents = await usecases.findLearningContentsByOrganizationId({ organizationId, locale });
-  return dependencies.frameworkWithoutSkillSerializer.serialize(learningContents);
+  return dependencies.frameworkWithSkillsSerializer.serialize(learningContents);
 };
 
 const targetProfileController = {
