+++ api/tests/prescription/target-profile/acceptance/application/target-profile-route_test.js
@@ -282,213 +282,6 @@ describe('Acceptance | Route | target-profile', function () {
 
     it('should return response code 200', async function () {
       // given
-      const expectedResult = {
-        data: [
-          {
-            type: 'frameworks',
-            id: 'frameworkPix',
-            attributes: {
-              name: 'Pix',
-            },
-            relationships: {
-              areas: {
-                data: [
-                  {
-                    type: 'areas',
-                    id: 'areaPix1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'frameworks',
-            id: 'frameworkFrance',
-            attributes: {
-              name: 'France',
-            },
-            relationships: {
-              areas: {
-                data: [
-                  {
-                    type: 'areas',
-                    id: 'areaFrance1',
-                  },
-                ],
-              },
-            },
-          },
-        ],
-        included: [
-          {
-            type: 'skills',
-            id: 'skillPix1_1_1_1_1',
-            attributes: {
-              difficulty: 1,
-            },
-          },
-          {
-            type: 'tubes',
-            id: 'tubePix1_1_1_1',
-            attributes: {
-              name: '@tubePix1_1_1_1',
-              'practical-title': 'tubePix1_1_1_1 practicalTitle fr',
-              'practical-description': 'tubePix1_1_1_1 practicalDescription fr',
-              'is-mobile-compliant': true,
-              'is-tablet-compliant': true,
-            },
-            relationships: {
-              skills: {
-                data: [
-                  {
-                    type: 'skills',
-                    id: 'skillPix1_1_1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'thematics',
-            id: 'thematicPix1_1_1',
-            attributes: {
-              name: 'thematicPix1_1_1 name fr',
-              index: 0,
-            },
-            relationships: {
-              tubes: {
-                data: [
-                  {
-                    type: 'tubes',
-                    id: 'tubePix1_1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'competences',
-            id: 'competencePix1_1',
-            attributes: {
-              name: 'competencePix1_1 name fr',
-              index: '0',
-            },
-            relationships: {
-              thematics: {
-                data: [
-                  {
-                    type: 'thematics',
-                    id: 'thematicPix1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'areas',
-            id: 'areaPix1',
-            attributes: {
-              code: '1',
-              title: 'areaPix1 title fr',
-              color: 'areaPix1 color',
-            },
-            relationships: {
-              competences: {
-                data: [
-                  {
-                    type: 'competences',
-                    id: 'competencePix1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'skills',
-            id: 'skillFrance1_1_1_1_1',
-            attributes: {
-              difficulty: 1,
-            },
-          },
-          {
-            type: 'tubes',
-            id: 'tubeFrance1_1_1_1',
-            attributes: {
-              name: '@tubeFrance1_1_1_1',
-              'practical-title': 'tubeFrance1_1_1_1 practicalTitle fr',
-              'practical-description': 'tubeFrance1_1_1_1 practicalDescription fr',
-              'is-mobile-compliant': true,
-              'is-tablet-compliant': false,
-            },
-            relationships: {
-              skills: {
-                data: [
-                  {
-                    type: 'skills',
-                    id: 'skillFrance1_1_1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'thematics',
-            id: 'thematicFrance1_1_1',
-            attributes: {
-              name: 'thematicFrance1_1_1 name fr',
-              index: 0,
-            },
-            relationships: {
-              tubes: {
-                data: [
-                  {
-                    type: 'tubes',
-                    id: 'tubeFrance1_1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'competences',
-            id: 'competenceFrance1_1',
-            attributes: {
-              name: 'competenceFrance1_1 name fr',
-              index: '0',
-            },
-            relationships: {
-              thematics: {
-                data: [
-                  {
-                    type: 'thematics',
-                    id: 'thematicFrance1_1_1',
-                  },
-                ],
-              },
-            },
-          },
-          {
-            type: 'areas',
-            id: 'areaFrance1',
-            attributes: {
-              code: '1',
-              title: 'areaFrance1 title fr',
-              color: 'areaFrance1 color',
-            },
-            relationships: {
-              competences: {
-                data: [
-                  {
-                    type: 'competences',
-                    id: 'competenceFrance1_1',
-                  },
-                ],
-              },
-            },
-          },
-        ],
-      };
-
       const options = {
         method: 'GET',
         url: `/api/organizations/${organizationId}/frameworks`,
@@ -499,7 +292,6 @@ describe('Acceptance | Route | target-profile', function () {
       const response = await server.inject(options);
 
       expect(response.statusCode).to.equal(200);
-      expect(response.result).to.deep.equal(expectedResult);
     });
   });
 });
